requirements.txt
astroid==2.3.3
autopep8==1.5
Click==7.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
hiredis==1.0.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.0.1.24
pathtools==0.1.2
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
python-dotenv==0.13.0
redis==3.4.1
redisearch @ git+https://github.com/quintonparker/redisearch-py.git@766d5cc9a02e6a11f87cab1f41bfb7976b5d02a7
rmtest==0.7.0
six==1.13.0
typed-ast==1.4.0
watchdog==0.10.2
Werkzeug==1.0.1
wrapt==1.11.2
